add `all` flag and new `get` command
Add new output formatting and commands to bpfctl. Specifically it adds
an `--all` flag to the list command which can be used to print ALL of the
programs on a node. Additionally with the new info and metadata provided
by bpfd we also add a new get command.  The get command can be used with
a kernel program ID to drill into very detailed information for a given
program.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>