cleanup image pulling
This commit removes the program overrides struct with the oci_utils
crate, and replaces it with simply the bytecode path and section name
since those are the only two fields that may be overridden by a bytecode
container image pull.

It also adds a new helper `data_mut()` to the `ProgramData` struct which
allows us to borrow the data mutably, updating the section name on
image pull if needed.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>