Support attaching uprobes to targets inside containers
This commit adds a new executable called bpfd-ns which can be called
to attach uprobes and uretprobes inside containers.  This executable
changes to the mount namespace of the container, does the attach, and
then changes back to the original mount namespace.

The executable is spawned from the uprobe code when a namespace_pid
parameter is provided when loading a uprobe.

Other changes:
- renamed namespace to namespace_pid in bpfd proto and make it an int32

Signed-off-by: Andre Fredette <afredette@redhat.com>