bpfman: re-write the database layer
This commit is huge and it re-writes the memory layout for all Program
related information. More specifically it removed the KernelInfo struct
and collapses all information for bpfman programs AND other programs
into a single Program enum. This program Enum is essentially a wrapper
for a sled database tree. Each program will get it's own sled database tree
and anytime a getter/setter is called on a Program it will interact with
the database.

On load we have to use a random number generator to generate a unique id
for the program tree prior to load.  Once the program is successfully
loaded we can then swap the root index of the database tree to the
correct id.  From then on it's always safe to assume the tree name will
correlate to a given program id.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>