remove lazy static root database
Now that bpfman is a library we need to make sure
that the sled database lock is released whenever
it's not being used so other processes can call the
library on the same host. This wan't possible with
the root database reference as a lazy static variable
since once it was initialized by a process it would
never be dropped until the calling process exited.
This is an issue when the calling process is our
long lived rpc shim.

To fix this simply add the root database instance
as a member of the BpfManager and then pass it via
arguments whereever it's needed.

Additionally fixup the retry mechanism for acquiring
the lock on the filesystem backed sled db instance
to default accordingly and tweak it to retry a bit
more when running in kubernetes.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>